interpipe: keep the appsrc byte level in a guint64
gst_app_src_get_current_level_bytes returns a guint64. Truncating it into a
guint makes a queued level that is an exact multiple of 2^32 read back as 0,
which releases a pending serial event before the buffers ahead of it have
drained.

Ported from https://github.com/cfsbhawkins/gst-interpipe (Brian Hawkins),
commit 9e78b96a.

+++ gst/interpipe/gstinterpipesrc.c
@@ -657,7 +657,7 @@ gst_inter_pipe_src_create (GstBaseSrc * base, guint64 offset, guint size,
   g_mutex_lock (&src->serial_events_lock);
   if (!g_queue_is_empty (src->pending_serial_events)) {
     GstEvent *head = g_queue_peek_head (src->pending_serial_events);
-    guint curr_bytes;
+    guint64 curr_bytes;
 
     GST_DEBUG_OBJECT (src,
         "Got event with timestamp %" GST_TIME_FORMAT,
